Kinda working but not really.
Fragments are now working, even reordering is working, but fragments do not get updated if changes were made on the other page.

For example: if I re-order books, the default author name on the next page will not change since it was already created and values were set. We are passing the strings, but should be somehow passing the reference to an object. Not sure how this can be achieved.