Processing books without TOC.
For the books which have no TOC or the TOC is empty we now create
an artificial TOC entry pointing to the first page of the book (which
is often cover page). Not doing this results in other books (with TOC)
being visible in resulting book table of contents, and books without
TOC missing there altogether.